Add files via upload
Ques. 20. Consider that a system has P resources of same type. These resources are shared by Q
processes time to time. All processes request and release the resources one at a time. Generate a
solution to demonstrate that, the system is in safe state when following conditions are satisfied.
Conditions:
1. Maximum resource need of each process is between 1 and P.
2. Summation of all maximum needs is less than P+Q